Minor refactor to initializing a logger for Reclass
The primary goal here was to drop the need/use of the get_options()
from reclass.config, and to instead rely on a parameter passed in from
the caller (for debug mode). This is meant to hold us over until the
config/option/cli refactor is complete - it will be easy to refactor
logging configuration later.